WIP: Global defaults for image configs
This is a first draft attempt to allow setting default image configs at
a global level.

There is one issue so far with respect to honoring defaults for the
`file` provider. Right now the implemenation for every other provider is
performed by passing defaults to `ValidateImage()`, however the file
provider does not use this.

Since the other providers identify config by iterating through and
parsing labels in `ValidateImage()`, the defaults can be checked there. The
file provider instead unmarshals the image config directly into the
`model.Image` struct. This poses a challenge for setting default values
for booleans as we have no way to detect the difference between an
absent or false value.

This could be mitigated by migrating the boolean values to pointers or
by using an intermediate struct which uses pointers that the file is
unmarshalled to and then coerced into the `model.Image` struct with
default values set.

Open to feedback on how to solve it but I'm leaning towards the
intermediate struct to avoid changing in as many places.

Fixes #491