Bump Go version up to 1.21
When the updated version of Nomad client comes out, I'm planning on a
patch to upgrade so that communication over unix socket is available,
however the latest requires go 1.21. Rather than bundle a go upgrade
with the Nomad change, I thought it best to split up and make sure any
potential issues or concerns with the go version are addressed first.